moved level token from pause to new script added urls for level weather stuff
