+++ Implementation/django_template/social_app/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
 from . import views, pause, friends, weather, gamestate, levelloader, invite, lobby, betweenlevels, steps, \
-    gravity_manager, collectables
+    gravity_manager, collectables, levelweather
 
 # TODO: @team add your source files here
 
@@ -55,7 +55,16 @@ urlpatterns = [
     path('store_current_weather/', weather.set_current_weather),
     path('load_tokens/', weather.get_weather_info),
     path('store_weather_info/', weather.update_player_weather),
-    path('set_level_current/', weather.set_current_weather),
+
+    # -------------------------------------------{Level Weather Stuff}-------------------------------------------------#
+    #  @Kerstin urls for level weather stuff [source: levelweather.py]
+    path('set_level_current/', levelweather.set_level_current),
+    path('get_level_current/', levelweather.get_level_current),
+    path('set_host_level_token/', levelweather.set_host_token),
+    path('set_joined_level_token/', levelweather.set_joined_token),
+    path('get_host_level_token/', levelweather.get_host_token),
+    path('get_joined_level_token/', levelweather.get_joined_token),
+    path('use_weather_token/', levelweather.use_level_token),
 
     # --------------------------------------------{Level Change Stuff}-------------------------------------------------#
     #  @Vyno urls for Level change [source: levelloader.py]
